v0.732
Updates:
- More work on Sandworm boss (not finished yet).

Adjustments:
- Increased the damage the Deer-Nymph takes from colliding with a Crystal Rock.
- Swapped out the Artifact Upgrade Help Menu sprites.

+++ StuckAtLv1/Assets/Scripts/Enemies/Common/Enemy.cs
@@ -350,7 +350,7 @@ Class that handles enemy stats and HP values and taking damage, as well as attac
         rend.color = new Color(1,1,1);
     }
 
-    public void SetSpeed(int speedToSet) {
+    public void SetSpeed(float speedToSet) {
         baseSpeed = speedToSet;
     }
